solve review comments
