+++ capitalizedecorator.rb
@@ -1,0 +1,6 @@
+require './decorator'
+class CapitalizeDecorator < Decorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
