+++ decorator.rb
@@ -1,6 +1,8 @@
 require './nameable'
 
 class Decorator < Nameable
+  attr_accessor :nameable
+
   def initialize(name)
     super()
     @nameable = name
@@ -8,17 +10,5 @@ class Decorator < Nameable
 
   def correct_name
     @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    @nameable.correct_name[0..9]
   end
 end
